// src/views/Products/SetPriceList/PriceList/CRUD/Create/utils/type.ts
export interface ICreateSettingPriceListType {
  status: number | null;
  priceName: string | null;
  applyStartDate: string | null;
  applyEndDate: string | null;
  note: string | null;
  store: number | null;
  channel: number | null;
  customerGroup: number | null;
  customerLevel: number | null;
  billCreator: string | null;
  newPriceSelect: number | null;
  newPrice: number | null;
  category: number | null;
}

export interface ISelectButtonType {
  add: boolean;
  deduction: boolean;
}

export interface ISelectUnitType {
  VND: boolean;
  percent: boolean;
}